C ffi用の関数定義を分離
voicevox_core crateとして Internalのstructがpublicとなっているのがおかしいのと、 `Internal` を `VoicevoxCore` に改名(Internalの機能こそがVoicevoxCoreの中心となるべき機能であると判断したため)
