+++ crates/voicevox_core/Cargo.toml
@@ -6,18 +6,13 @@ edition = "2021"
 [features]
 default = []
 directml = ["onnxruntime/directml"]
-generate-c-header = ["dep:cbindgen"]
 
-[lib]
-name = "core"
-crate-type = ["cdylib"]
 
 [dependencies]
 anyhow = "1.0.61"
 cfg-if = "1.0.0"
 derive-getters = "0.2.0"
 derive-new = "0.5.9"
-libc = "0.2.131"
 once_cell = "1.10.0"
 onnxruntime = { git = "https://github.com/VOICEVOX/onnxruntime-rs.git", version = "0.1.0" }
 serde = { version = "1.0.143", features = ["derive"] }
@@ -33,7 +28,3 @@ async-std = { version = "1.12.0", features = ["attributes"] }
 surf = "2.3.2"
 flate2 = "1.0.24"
 tar = "0.4.38"
-
-
-[build-dependencies]
-cbindgen = { version = "0.24.3", optional = true }
