+++ crates/voicevox_core/src/result_code.rs
@@ -1,0 +1,21 @@
+#[repr(i32)]
+#[derive(Debug, PartialEq, Eq)]
+#[allow(non_camel_case_types)]
+pub enum VoicevoxResultCode {
+    // C でのenum定義に合わせて大文字で定義している
+    // 出力フォーマットを変更すればRustでよく使われているUpperCamelにできるが、実際に出力されるコードとの差異をできるだけ少なくするため
+    VOICEVOX_RESULT_SUCCEED = 0,
+    VOICEVOX_RESULT_NOT_LOADED_OPENJTALK_DICT = 1,
+    VOICEVOX_RESULT_FAILED_LOAD_MODEL = 2,
+    VOICEVOX_RESULT_FAILED_GET_SUPPORTED_DEVICES = 3,
+    VOICEVOX_RESULT_CANT_GPU_SUPPORT = 4,
+    VOICEVOX_RESULT_FAILED_LOAD_METAS = 5,
+    VOICEVOX_RESULT_UNINITIALIZED_STATUS = 6,
+    VOICEVOX_RESULT_INVALID_SPEAKER_ID = 7,
+    VOICEVOX_RESULT_INVALID_MODEL_INDEX = 8,
+    VOICEVOX_RESULT_INFERENCE_FAILED = 9,
+    VOICEVOX_RESULT_FAILED_EXTRACT_FULL_CONTEXT_LABEL = 10,
+    VOICEVOX_RESULT_INVALID_UTF8_INPUT = 11,
+    VOICEVOX_RESULT_FAILED_PARSE_KANA = 12,
+    VOICEVOX_RESULT_INVALID_AUDIO_QUERY = 13,
+}
