+++ crates/voicevox_core_c_ffi/Cargo.toml
@@ -1,0 +1,26 @@
+[package]
+name = "voicevox_core_c_ffi"
+version = "0.1.0"
+edition = "2021"
+
+[lib]
+name = "core"
+crate-type = ["cdylib"]
+
+[features]
+generate-c-header = ["dep:cbindgen"]
+directml = ["voicevox_core/directml"]
+
+[dependencies]
+voicevox_core = { path = "../voicevox_core" }
+libc = "0.2.131"
+once_cell = "1.10.0"
+serde_json = "1.0.83"
+
+[dev-dependencies]
+pretty_assertions = "1.2.1"
+anyhow = "1.0.61"
+rstest = "0.15.0"
+
+[build-dependencies]
+cbindgen = { version = "0.24.3", optional = true }
